Move kubectl e2e tests to their own directory and prefix the test names with [sig-cli]
