+++ test/e2e/BUILD
@@ -73,7 +73,6 @@ go_library(
         "ingress.go",
         "kibana_logging.go",
         "kube_proxy.go",
-        "kubectl.go",
         "kubelet.go",
         "kubelet_perf.go",
         "limit_range.go",
@@ -90,9 +89,7 @@ go_library(
         "pod_gc.go",
         "podpreset.go",
         "pods.go",
-        "portforward.go",
         "pre_stop.go",
-        "proxy.go",
         "reboot.go",
         "resize_nodes.go",
         "resource_quota.go",
@@ -112,7 +109,6 @@ go_library(
         "//pkg/api:go_default_library",
         "//pkg/api/v1/pod:go_default_library",
         "//pkg/api/v1/service:go_default_library",
-        "//pkg/apis/batch/v2alpha1:go_default_library",
         "//pkg/apis/networking:go_default_library",
         "//pkg/client/clientset_generated/clientset:go_default_library",
         "//pkg/client/clientset_generated/clientset/typed/certificates/v1beta1:go_default_library",
@@ -123,14 +119,12 @@ go_library(
         "//pkg/controller:go_default_library",
         "//pkg/controller/endpoint:go_default_library",
         "//pkg/controller/node:go_default_library",
-        "//pkg/kubectl/cmd/util:go_default_library",
         "//pkg/kubelet/apis:go_default_library",
         "//pkg/kubelet/apis/stats/v1alpha1:go_default_library",
         "//pkg/kubelet/pod:go_default_library",
         "//pkg/master/ports:go_default_library",
         "//pkg/metrics:go_default_library",
         "//pkg/quota/evaluator/core:go_default_library",
-        "//pkg/util/exec:go_default_library",
         "//pkg/util/logs:go_default_library",
         "//pkg/util/version:go_default_library",
         "//plugin/pkg/admission/serviceaccount:go_default_library",
@@ -141,15 +135,12 @@ go_library(
         "//test/e2e/generated:go_default_library",
         "//test/e2e/manifest:go_default_library",
         "//test/e2e/perf:go_default_library",
-        "//test/e2e/scheduling:go_default_library",
         "//test/e2e/upgrades:go_default_library",
         "//test/e2e/workload:go_default_library",
         "//test/e2e_federation:go_default_library",
         "//test/images/net/nat:go_default_library",
         "//test/utils:go_default_library",
         "//test/utils/junit:go_default_library",
-        "//vendor/github.com/elazarl/goproxy:go_default_library",
-        "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/google/cadvisor/info/v1:go_default_library",
         "//vendor/github.com/influxdata/influxdb/client/v2:go_default_library",
@@ -158,7 +149,6 @@ go_library(
         "//vendor/github.com/onsi/ginkgo/reporters:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/golang.org/x/crypto/ssh:go_default_library",
-        "//vendor/golang.org/x/net/websocket:go_default_library",
         "//vendor/golang.org/x/oauth2/google:go_default_library",
         "//vendor/google.golang.org/api/monitoring/v3:go_default_library",
         "//vendor/k8s.io/api/batch/v1:go_default_library",
@@ -181,7 +171,6 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/runtime/schema:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/intstr:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/util/net:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/runtime:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/util/uuid:go_default_library",
@@ -189,7 +178,6 @@ go_library(
         "//vendor/k8s.io/apimachinery/pkg/util/yaml:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/watch:go_default_library",
         "//vendor/k8s.io/apiserver/pkg/authentication/serviceaccount:go_default_library",
-        "//vendor/k8s.io/apiserver/pkg/registry/generic/registry:go_default_library",
         "//vendor/k8s.io/client-go/discovery:go_default_library",
         "//vendor/k8s.io/client-go/kubernetes:go_default_library",
         "//vendor/k8s.io/client-go/rest:go_default_library",
@@ -227,6 +215,7 @@ filegroup(
         "//test/e2e/extension:all-srcs",
         "//test/e2e/framework:all-srcs",
         "//test/e2e/generated:all-srcs",
+        "//test/e2e/kubectl:all-srcs",
         "//test/e2e/manifest:all-srcs",
         "//test/e2e/perf:all-srcs",
         "//test/e2e/perftype:all-srcs",
