+++ test/e2e/kubectl/OWNERS
@@ -1,0 +1,4 @@
+approvers:
+- sig-cli-maintainers
+reviewers:
+- sig-cli
